refactor : divide userlist into list and item component
